pyproject.toml
[tool.poetry]
name = "the-project-tracker"
version = "0.1.0"
description = "Code projects AI tracker"
authors = ["Carlos Vecina <carlos.vecina@jobandtalent.com>"]
readme = "README.md"
packages = [{include = "the_project_tracker"}]

[tool.poetry.dependencies]
python = "^3.11"
selenium = "^4.8.2"
pydantic = "^1.10.6"
SQLAlchemy = "^2.0.6"
pandas = "^1.5.3"
pytest = "^7.2.2"
ipykernel = "^6.21.3"
requests = "^2.28.2"
click = "^8.1.3"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
